+++ tests/test_core.py
@@ -1,3 +1,4 @@
+import platform
 from time import sleep
 from zipfile import ZipFile, ZipInfo
 
@@ -8,11 +9,12 @@ from tests.utils import (
     dir_tree_factory,
     file_factory,
     hash_file,
+    umask,
 )
 
 
-def test_write_dir_tree(base_path):
-    """Archiving a directory tree works."""
+def test_write_dir_tree_mtime(base_path):
+    """Archiving a directory tree works with different modified time."""
     dir_tree = dir_tree_factory(base_path)
 
     # Create base ReproducibleZipFile archive
@@ -27,7 +29,7 @@ def test_write_dir_tree(base_path):
         for path in sorted(dir_tree.glob("**/*")):
             zp.write(path)
 
-    # Update modified times
+    # Sleep to update modified times, change permissions
     sleep(2)
     for path in dir_tree.glob("**/*"):
         path.touch()
@@ -52,6 +54,56 @@ def test_write_dir_tree(base_path):
     # ReproducibleZipFile hashes should match; ZipFile hashes should not
     assert hash_file(repro_zipfile_arc1) == hash_file(repro_zipfile_arc2)
     assert hash_file(zipfile_arc1) != hash_file(zipfile_arc2)
+
+
+def test_write_dir_tree_mode(base_path):
+    """Archiving a directory tree works with different permission modes."""
+    with umask(0o022):
+        dir_tree = dir_tree_factory(base_path)
+
+    # Create base ReproducibleZipFile archive
+    repro_zipfile_arc1 = base_path / "repro_zipfile1.zip"
+    with ReproducibleZipFile(repro_zipfile_arc1, "w") as zp:
+        for path in sorted(dir_tree.glob("**/*")):
+            zp.write(path)
+
+    # Create regular ZipFile archive for comparison
+    zipfile_arc1 = base_path / "zipfile1.zip"
+    with ZipFile(zipfile_arc1, "w") as zp:
+        for path in sorted(dir_tree.glob("**/*")):
+            zp.write(path)
+
+    # Change permissions
+    with umask(0o002) as mask:
+        dir_tree.chmod(mode=0o777 ^ mask)
+        for path in dir_tree.glob("**/*"):
+            if path.is_file():
+                path.chmod(mode=0o666 ^ mask)
+            else:
+                path.chmod(mode=0o777 ^ mask)
+
+    # Create second ReproducibleZipFile archive after delay
+    repro_zipfile_arc2 = base_path / "repro_zipfile2.zip"
+    with ReproducibleZipFile(repro_zipfile_arc2, "w") as zp:
+        for path in sorted(dir_tree.glob("**/*")):
+            zp.write(path)
+
+    # Create second regular ZipFile archive for comparison after delay
+    zipfile_arc2 = base_path / "zipfile2.zip"
+    with ZipFile(zipfile_arc2, "w") as zp:
+        for path in sorted(dir_tree.glob("**/*")):
+            zp.write(path)
+
+    # All four archives should have identical content
+    assert_archive_contents_equals(repro_zipfile_arc1, zipfile_arc1)
+    assert_archive_contents_equals(repro_zipfile_arc1, repro_zipfile_arc2)
+    assert_archive_contents_equals(repro_zipfile_arc1, zipfile_arc2)
+
+    # ReproducibleZipFile hashes should match; ZipFile hashes should not
+    assert hash_file(repro_zipfile_arc1) == hash_file(repro_zipfile_arc2)
+    if platform.system() != "Windows":
+        # Windows doesn't seem to actually make them different
+        assert hash_file(zipfile_arc1) != hash_file(zipfile_arc2)
 
 
 def test_write_dir_tree_string_paths(rel_path):
@@ -132,7 +184,7 @@ def test_write_single_file(base_path):
     assert hash_file(zip1) != hash_file(zip2)
 
 
-def test_write_single_file_with_source_date_epoch(base_path, monkeypatch):
+def test_write_single_file_source_date_epoch(base_path, monkeypatch):
     """Writing the same file with different mtime with SOURCE_DATE_EPOCH set produces the
     same hash."""
 
@@ -163,6 +215,26 @@ def test_write_single_file_with_source_date_epoch(base_path, monkeypatch):
     # Base archive hash should match neither, two archives with SOURCE_DATE_EPOCH should match
     assert hash_file(arc_base) != hash_file(arc_sde1)
     assert hash_file(arc_sde1) == hash_file(arc_sde2)
+
+
+def test_write_single_file_file_mode_env_var(rel_path, monkeypatch):
+    """REPRO_ZIPFILE_FILE_MODE environment variable works."""
+
+    with umask(0o002):
+        # Expect 664
+        data_file = file_factory(rel_path)
+
+    monkeypatch.setenv("REPRO_ZIPFILE_FILE_MODE", "600")  # rw-------
+
+    arc_path = rel_path / "archive.zip"
+    with ReproducibleZipFile(arc_path, "w") as zp:
+        zp.write(data_file)
+
+    with ZipFile(arc_path, "r") as zp:
+        print(zp.infolist())
+        mode = (zp.getinfo(data_file.name).external_attr >> 16) & 0o777
+
+    assert mode == 0o600, (oct(mode), oct(0o600))
 
 
 def test_write_single_file_string_paths(rel_path):
@@ -232,6 +304,27 @@ def test_write_single_file_arcname(base_path):
     # ReproducibleZipFile hashes should match; ZipFile hashes should not
     assert hash_file(repro_zip1) == hash_file(repro_zip2)
     assert hash_file(zip1) != hash_file(zip2)
+
+
+def test_write_single_dir_dir_mode_env_var(rel_path, monkeypatch):
+    """REPRO_ZIPFILE_DIR_MODE environment variable works."""
+
+    with umask(0o002):
+        # Expect 775
+        dir_path = rel_path / data_factory()
+        dir_path.mkdir()
+
+    monkeypatch.setenv("REPRO_ZIPFILE_DIR_MODE", "700")  # rwx------
+
+    arc_path = rel_path / "archive.zip"
+    with ReproducibleZipFile(arc_path, "w") as zp:
+        zp.write(dir_path)
+
+    with ZipFile(arc_path, "r") as zp:
+        print(zp.infolist())
+        mode = (zp.getinfo(dir_path.name + "/").external_attr >> 16) & 0o777
+
+    assert mode == 0o700, (oct(mode), oct(0o700))
 
 
 def test_writestr(tmp_path):